Improved alarm, date, and time displays
Before setting the time, the clock prompts for a selection between
"24-hour" and "12-hour".  Before setting the date, the clock prompts
for a selection between "dd-mm-yy" (EU format) and "mm-dd-yy"
(US format).

While the clock is displaying the time, the user may see the current
day of week, month, and day of month by pressing the set or plus
buttons (like the original Adafruit firmware).

If in the time display mode, the clock displays an alarm status
message when the alarm is turned on and off (like the original
Adafruit firmware).

The display_string("...") function has been removed and replaced by
display_pstr(PSTR("...")), which uses program memory for the string.