Pins optimized for clock sleep
The unconnected pins now have pullups enabled, and the button and
alarm switch pins are clamped to ground during sleep.  The sleep
switch might be shorted to ground and, if the clock is tightly packed
in a suitcase, so might the buttons.  So for the button and alarm
switch pins, clamping to ground seems preferable to enabling pull-up
resistors.  The remain pins have been similarly optimized while in
sleep mode.

Oddly enough, I do not observe any power savings as a result of these
efforts!  Oh well...  I suppose doing the theoretically right thing is
worth committing, as long as the changes do no harm.