Added autodim support
With a photoresistor and 10K pull-up installed in the expansion holes
of an Ice Tube Clock, display brightness can be automatically
controlled by ambient light.

Because I blew my atmega128v, I'm running on an atmega328p.  Sleep now
seems to work intermittently, and I do not know if this is do to the
hardware change, this update, or is just something I never noticed
before.