Brightness control improvements
The per digit brightness control no longer flickers.  Before the
per-digit brightness control would flicker when a user increased
around 1/2 of the digits to maximum brightness---not a typical usage
scenario but might happen nonetheless.

The autodim minimum and maximum values may now range from -5 to
20---beyond the allowed brightness range of 0 to 10.  The actual
display brightness, however, will never be brighter or dimmer than
allowed to ensure that the display will always be visible and prevent
display damage.

With the autodim feature, display brightness is calculated from
photoresistor voltage.  In the dark, voltage is highest and decreases
as the ambient light increases.  The minimum and maximum autodim
values set the brightness at 5 volts and ground level.  Display
brightness is linearly interpolated from the photoresistor voltage.
If the interpolation yields a value above or below the allowed
brightness, display brightness is changed to the minimum or maximum
allowed brightness, respectively.

In normal clock use, photoresistor voltages change over only part of
the 0 to 5 volt range.  So setting the autodim thresholds below and
above the minimum and maximum provides more flexibility when
configuring autodimming.