Added GPS support
A GPS configured to output data at 9600 baud may be connected to the
ATmega RXD pin.  The RMC NMEA sentences are parsed and used to
automatically set the current time.

When the firmware receives GPS data and the GPS has a good signal,
the "set time", and "set date" menus disappear.  In their place is the
"set zone" menu, allowing the user to set the time difference between
local standard time and UTC/GMT.  Because the GPS outputs time in
UTC/GMT, the time difference is necessary for computing local time.

When the firmware receives GPS data but the GPS does not have a
signal, the clock alternates between the current time and "gps lost".

When the firmware is stops receiving GPS data or the GPS does not have
a signal, the "set zone" menu is replaced by the "set time" and
"set date" menus.

This commit also includes a more intuitive interface for setting the date.