Multiple alarms by days-of-week
Three alarms are now supported, but more can be added at compile time
by changing the ALARM_COUNT macro in alarm.h to something more than 3.

Each alarm may be configured to sound on "all days", "weekdays",
"weekends", or "custom".  The last option allows users to set a custom
subset of days-of-week for the particular alarm.  Naturally, each
alarm may be enabled or disabled through the menu, so one need not set
all three alarms.  When the alarm switch is on, only enabled alarms
will sound.  When the alarm switch is off, the alarm will never sound.