Reorganized menus using a nested structure
The menu structure is now nested allowing for better organization.

A major change is that the menu button now toggles between the clock
display and the menu structure.  The "+" button is used for moving to
the next item.  This makes the meaning of the "+" button consistent
across the interface, but is a major difference from the original
Adafruit firmware.