Improved Adafruit-style buttons
The compile-time option for Adafruit-style buttons is now in the
Makefile (config.h no longer exists), and this option is now enabled
by default.  Although the new button scheme is more logical, anyone
trying this firmware will already be familiar with the old Adafruit
scheme, so I think it makes more sense to use that as the default.

Defining the "ADAFRUIT_BUTTONS" macros also disables the circular
menus for a more Adafruit-style behavior.  In other words, continually
pressing menu while in the menus will eventually move to the
time-display screen.