DS18B20 Temperature sensor support
This is currently a failed hack, but I wanted to keep the code anyway
because I may be able to salvage it in the future.  So I'm creating a
separate branch named temperature to store this work.

The DS18B20 is installed by connecting the GND, DQ, and Vdd pins to
the ATmega's GND, PC2, and PC5 pins with a 4.7k pull-up between PC2
and PC5.  I drilled a small hole in the base of the case and inserted
the head of the DS18B20 through the hole to get ambient temperature.

This commit displays temperature (in Fahrenheit) along with the date.

I consider this a failed hack because the Ice Tube Clock generates
heat, and the clock itself and surrounding areas has a temperature
that is 1-8 degrees Fahrenheit higher than ambient temperature.  So,
the temperature reading is seldom accurate.  :-(

This commit does include code that is a good example of how to query
the DS18B20 via the 1-Wire protocol.