Added alternative time formats
This commit includes support for alternative time formats as well as a
couple more date formats.  Alternative time formats, including 12/24
hour mode, may be configured under "set regn"/"time fmt".

The user may also choose to disable display of daylight saving time.
DST, when displayed, is indicated by the circle on the 0th character
or, if the circle of the 0th character is needed to indicate AM/PM,
DST is indicated by the decimal of the 8th character.

Finally, numbers with leading zeros have been reformatted to omit
those zeros.