Extended auto-off features to extend display life
Three new auto-off options allow users to disable the display on a
regular schedule to extend the VFD tube lifetime.

The existing option to disable the display at night (when dark) is now
configured under "cfg disp"/"auto off"/"off dark".

Under "cfg disp"/"auto off"/"off time", the display may now be
disabled for a given time period every day.  For example, if the clock
is at the workplace of someone with a 9-to-5 schedule, the display may
be disabled from 5:00pm to 9:00am.

Under "cfg disp"/"auto off"/"off days", the display may be disabled
for specific days each week.  For example, if the clock is at the
workplace of someone with a Monday-through-Friday schedule, the
display may be disabled during weekends.

Under "cfg disp"/"auto off"/"on days", the display may be enabled for
specific days each way, overriding the aforementioned "off time"
setting.  For example, if the clock is at the home of a person with a
9-to-5 Monday-through-Friday schedule, they might set the "off time"
to 9:00am to 5:00 pm and set the "on days" to weekends, so the display
would be enabled whenever the person is home.  Note that "on days"
does not override the "off dark" settings.