Minor bugfixes; Dragon compatibility; updated docs
An alarm volume of 10 saved to EEPROM would be restored as a volume
of 0!  This only happened after a microcontroller reset which only
happens if the clock battery is absent or dead.  This has been fixed.

The controls in config.h to compensate for phosphor loss in an aging
VFD tube did not work at all!  This has been fixed.

The Makefile has been modified to allow compatibility with the
Atmel AVR Dragon programmer.

The default alarm volume is now a progressive alarm from volume 0 to
volume 10 over a 1 minute period.

Documentation in the README and config.h has been updated.