Extended battery hack deprecated
After the previous commit, power used by an Ice Tube Clock in the
without any hardware modification was 10.6 uA.

To make the clock more efficient, I tried cutting the trace between
the resistor divider and system power, instead connecting the resistor
divider directly to the voltage divider.  This setups is just like the
extended battery hack except there is no second wire soldered between
the voltage regulator and PC1.  After that change current at 3.3v fell
to 1.5 uA--better than the extended battery hack!

So there is no longer a need for a software component of the extended
battery hack.  The PC1 is tri-stated so future versions of this
firmware will be backwards compatible with hardware modified for the
extended battery hack