First EEPROM byte now reserved and ignored
According to William Phelps (wbp), the first byte of EEPROM in ATMEGA
chips is unreliable in certain circumstances:

  http://forums.adafruit.com/viewtopic.php?f=41&t=40905

I've never noticed this problem, but it is possible that others have.

To eliminate this potential problem, I now allocate a bogus EEPROM
variable in icetube.c, and the variable is never used.  Because
icetube.o is the first object given to the linker, this bogus variable
is allocated as the first memory byte.  The end result is that the
first EEPROM byte is no longer used for storing data.