Strange MCUSR behavior workaround
When the microcontroller starts, this firmware checks the MCUSR
register to determine the reason for the reset and presents a suitable
diagnostic message on the display.

For whatever reason, the BORF flag is sometimes set in addition to
other reset flags when no brown out condition has occurred.  The brown
out detection reset message ("bod rset") is now displayed only if no
other reset flags are set.