Minor time format menu bug fixed
For the 12-hour time formats, the PM indicator (leftmost dot) should
only indicate PM for time formats that do not already display AM or PM
after the time.

In some circumstances, it was possible to enable the PM indicator,
even if AM or PM was also displayed after the time.  This bug in the
configuration menu has been fixed.