Aesthetic bugfix to hour-minute time display
The hours are now displayed starting on the second and third display
digits for the hour-minute time display format which displays only the
hour and the minute.  This fixes a minor bug introduced a few commits
ago which caused the hours to be displayed on the first and second
display positions.  The hour-minute time display format looks better
now.