xmas-icetube board (rev. D) documentation updates
My Rev. D boards arrived yesterday evening.  I've built and tested
that board and everything is working wonderfully!

This commit moves both the case and board design files into the same
directory.  The corresponding READMEs have been merged, updated for
the Rev D. main board, and now include sections for finding the parts
and building the clock.  Also the manifest lists part manufacturers,
and the case design is clearly marked as "REV A."