Added cathode driver compile-time options
When using the to-spec hack or the xmas-icetube hardware revision, the
filament can be driven with direct current and/or at reduced voltage
(~3.3v or ~2.5v).  Although these options are provided for testing
purposes, I can imagine that one might want to reduce the voltage
across the filament to reduce the reddish glow, although doing so may
reduce tube life.