Improved initial default settings
The default time format, date format, timezone offset, and DST
settings now default to the computer on which you compile this
firmware.  Also automatic dimming is enabled by default, if that when
that feature is available.  Display pulsing during alarm and snooze is
now also enabled by default.

This commit also fixes a bug which caused the automatic fall DST
correction in the European Union to happen an hour ahead of schedule.