Increased GPS synchronization accuracy
With the GPS mod, the clock now synchronizes time to within a half
second of the GPS-reported time, as opposed to 1.5 seconds before.

The disadvantage of this change is that automatic drift correction
statistics are not collected while receiving GPS time data.  The many
corrections required to synchronize time with subsecond accuracy would
introduce too much rounding error.  But those with a GPS module should
not be too concerned with timekeeping drift correction anyway.