External TCXO & sleep bugfix
With an external clock such as the DS32kHz TCXO, voltage would not
stabilize sufficiently for the repeated analog-to-digital measurements
of battery voltage during sleep to converge.  As a result, the
analog-to-digital converter would run continuously and drain the
battery in a short amount of time--around thirty minutes.

To resolve the issue, the voltage measurement algorithm now uses a
looser criteria for convergence and enforces a maximum number of
analog-to-digital conversions.