Drift correction algorithm now tunable in config.h
A drift correction constant may now be specified at compile time in
two ways.

First, the drift correction value may be preloaded.  In this case,
the clock will still calculate how fast or slow it's running when the
user corrects the time.  This will allow the clock to refine the
preloaded drift correction value and automatically adapt to frequency
changes as the crystal ages.

Second, the drift correction value may be defined as a constant, in
which case the clock will always use the drift correction value
defined at compile time.  This feature also allows drift correction to
be disabled, as a drift correction value of zero means that no drift
correction should be performed.

The comments in config.h describe how to use these options.