Temperature sensor errors are now displayed
If the TEMPERATURE_SENSOR macro is enabled in config.h, the clock will
use an attached DS18B20 temperature sensor to perform software
temperature compensation.

But if communication with the sensor failed, the clock would appear to
function normally--except that no temperature compensation would be
performed.  Now, the clock will periodically display a "temp err"
message to show the communication failure.

The absence of the "temp err" message will now provide assurance that
the clock is communicating with the DS18B20 and that temperature
compensation is being performed.