Documented diagnostic messages (thanks Doug!)
At the suggestion of Douglas Johnson, the troubleshooting section in
the firmware/README now contains a description of the various
diagnostic messages that might flash on the display.

The "ext rset" message has also been changed to "pin rset" due to the
difficulty of displaying an "x" on a 7-segment display.