Changed initial defaults in config.h
The defaults in config.h have been changed to increase the display
brightness range, allowing the user to increase the brightness of an
aging display through the menus without the need to reconfigure and
reprogram the firmware.  (firmware/config.h)

For the xmas hardware revision, the initial/preloaded drift correction
constant is now set by default to a value calibrated with the
recommended parts on a OSH Park board.  (hardware/config.h)