Minor time format bugfix
The AM/PM indicator (circle indicator) should be disabled when the
time format explicitly displays the AM/PM status as letters and should
be enabled in time format displays that do not.

Due to a menu bug in mode.c, the AM/PM indicator could occasionally
enabled or disabled arbitrarily when changing the time format in the
"cfg regn" / "time fmt" menu.

This has been fixed.