Added ability to disable "gps lost" message
The "gps lost" message normally flashes when the clock is receiving
data from an attached GPS unit, but the GPS unit has been unable to
achieve a lock for three minutes.

Occasionally, a user might wish to disable this message; this is now
possible as a compile-time option in config.h.