Updated code for avr-gcc (9.1.0, 2019)
The xmas code is now compatible with avr-gcc 9.1.0.  (I am using
Windows x64 binaries from http://blog.zakkemble.net/avr-gcc-builds/ .)
Three changes were required to support this kit:

(1) All inline functions are now declared as static to comply with
better enforcement of the C99 standard.

(2) Compiler optimization was interfering with the timed register
writes required to disable BOD during sleep.  Sleep worked fine, but
the ATmega328p was using ~20 uA during sleep rather than ~2 uA.
Moving the relevant sleep code to a separate function seems to fix the
issue.

(3) The newer version of avrdude requires reserved fuse bits to be set
whereas older versions requires them to be cleared.  This is now
configurable in the Makefile to support both ways.  See the FUSE_OPT
macro.

As before, the code continues to work with the avr-gcc version that
currently comes with Crosspack for MacOS X (avr-gcc 4.8.1, 2013) and
with WinAVR for Windows (avr-gcc 4.4.3, 2010).  It's hard to believe
that WinAVR is almost a decade old!